+++ R/feature_engineering.R
@@ -115,8 +115,11 @@ crf_cbind_attributes_single <- function(data, field, by, from = -2, to = 2, ngra
 #' @title Extract basic text features which are useful for entity recognition
 #' @description Extract basic text features which are useful for entity recognition
 #' @param x a character vector
-#' @param type a character string, which can be one of 'is_capitalised', 'is_url', 'is_email', 'is_number'
-#' @return a logical vector of the same length as \code{x}, indicating if \code{x} is capitalised, a url, an email or a number
+#' @param type a character string, which can be one of 'is_capitalised', 'is_url', 'is_email', 'is_number', 'prefix', 'suffix'
+#' @param n for type 'prefix' or 'suffix', the number of characters of the prefix/suffix
+#' @return 
+#' For type 'is_capitalised', 'is_url', 'is_email', 'is_number': a logical vector of the same length as \code{x}, indicating if \code{x} is capitalised, a url, an email or a number\cr
+#' For type 'prefix', 'suffix': a character vector of the same length as \code{x}, containing the prefix or suffix \code{n} number of characters of \code{x}
 #' @export
 #' @examples 
 #' txt_feature("Red Devils", type = "is_capitalised")
@@ -128,7 +131,9 @@ crf_cbind_attributes_single <- function(data, field, by, from = -2, to = 2, ngra
 #' txt_feature("123.15", type = "is_number")
 #' txt_feature("123,15", type = "is_number")
 #' txt_feature("123abc", type = "is_number")
-txt_feature <- function(x, type = c('is_capitalised', 'is_url', 'is_email', 'is_number')){
+#' txt_feature("abcdefghijklmnopqrstuvwxyz", type = "prefix", n = 3)
+#' txt_feature("abcdefghijklmnopqrstuvwxyz", type = "suffix", n = 3)
+txt_feature <- function(x, type = c('is_capitalised', 'is_url', 'is_email', 'is_number', 'prefix', 'suffix'), n = 4){
   type <- match.arg(type)
   missing <- is.na(x)
   miss <- as.logical(NA)
@@ -140,6 +145,12 @@ txt_feature <- function(x, type = c('is_capitalised', 'is_url', 'is_email', 'is_
     result <- grepl(".+@.+\\.", x)
   }else if(type == "is_number"){
     result <- grepl("^[[:digit:].,]+$", x)
+  }else if(type == "prefix"){
+    result <- substr(x, start = 1, stop = n)
+  }else if(type == "suffix"){
+    x_size <- nchar(x)
+    start <- x_size - (n - 1)
+    result <- substr(x, start = ifelse(start < 1, 1, start), stop = x_size)
   }
   result[missing] <- miss
   result